Refactor mock HTTP server into its own function
