+++ samples/ios/app/glean-sample-appUITests/ViewControllerTest.swift
@@ -23,34 +23,12 @@ class ViewControllerTest: XCTestCase {
         app = XCUIApplication()
     }
 
-    override func tearDown() {
-        server?.stop()
-    }
-
-    func setupServer(expectPingType: String) {
-        server = HttpServer()
-
-        server!["/submit/:appid/:ping/:schema/:pinguuid"] = { request in
-            let pingName = request.params[":ping"]!
-            if pingName == expectPingType {
-                let body = String(bytes: request.body, encoding: .utf8)!
-                let data = body.data(using: .utf8)!
-                print("Received data: \(body)")
-                let json = try! JSONSerialization.jsonObject(with: data, options: []) as? [String: Any]
-                self.lastPingJson = json
-
-                // Fulfill test's expectation once we parsed the incoming data.
-                self.expectation?.fulfill()
-            }
-            return HttpResponse.ok(.text("OK"))
+    func setupServer(expectPingType: String) -> HttpServer {
+        return mockServer(expectPingType: expectPingType) { json in
+            self.lastPingJson = json
+            // Fulfill test's expectation once we parsed the incoming data.
+            self.expectation?.fulfill()
         }
-        // For logging purposes:
-        server!.middleware.append { request in
-            print("Middleware: \(request.address ?? "unknown address") -> \(request.method) -> \(request.path)")
-            return nil
-        }
-
-        try! server!.start(9080)
     }
 
     func checkCustomCounterData(expectedValue: UInt64) {
@@ -61,11 +39,10 @@ class ViewControllerTest: XCTestCase {
         let counters = metrics["counter"] as! [String: Any]
         let value = counters["custom.counter"] as! UInt64
         XCTAssertEqual(value, expectedValue)
-
     }
 
     func testViewControllerInteraction() {
-        setupServer(expectPingType: "sample")
+        let server = setupServer(expectPingType: "sample")
 
         app.launchArguments = ["USE_MOCK_SERVER"]
         app.launch()
@@ -97,5 +74,7 @@ class ViewControllerTest: XCTestCase {
         }
 
         checkCustomCounterData(expectedValue: 4)
+
+        server.stop()
     }
 }
